remove vsl separation and add dynamic aws form
