+++ src/app/storage/components/AddEditStorageModal/AddEditStorageForm.tsx
@@ -13,9 +13,6 @@ import {
 } from '@patternfly/react-core';
 import Select from 'react-select';
 import { withFormik, FormikProps } from 'formik';
-import FormErrorDiv from '../../../common/components/FormErrorDiv';
-import KeyDisplayIcon from '../../../common/components/KeyDisplayIcon';
-import HideWrapper from '../../../common/components/HideWrapper';
 import utils from '../../../common/duck/utils';
 import storageUtils from '../../duck/utils';
 import commonUtils from '../../../common/duck/utils';
@@ -29,15 +26,8 @@ import {
 import ConnectionStatusLabel from '../../../common/components/ConnectionStatusLabel';
 import { OutlinedQuestionCircleIcon } from '@patternfly/react-icons';
 import { Box, Flex, Text } from '@rebass/emotion';
+import AWSForm from './ProviderForms/AWSForm';
 
-const nameKey = 'name';
-const s3UrlKey = 's3Url';
-const bucketNameKey = 'bucketName';
-const bucketRegionKey = 'bucketRegion';
-const accessKeyKey = 'accessKey';
-const secretKey = 'secret';
-const vslConfigKey = 'vslConfig';
-const vslBlobKey = 'vslBlob';
 
 const componentTypeStr = 'Repository';
 const currentStatusFn = addEditStatusText(componentTypeStr);
@@ -86,8 +76,6 @@ const InnerAddEditStorageForm = (props: IOtherProps & FormikProps<IFormValues>) 
   } = props;
 
 
-  const [isAccessKeyHidden, setIsAccessKeyHidden] = useState(true);
-  const [isSharedCred, setIsSharedCred] = useState(true);
   const [selectedProvider, setSelectedProvider] = useState(null);
   const [providerOptions, setproviderOptions] = useState([
     { label: 'aws', value: 'aws' },
@@ -95,25 +83,11 @@ const InnerAddEditStorageForm = (props: IOtherProps & FormikProps<IFormValues>) 
     { label: 'azure', value: 'azure' }
   ]);
 
-  const handleAccessKeyHiddenToggle = e => {
-    e.preventDefault();
-    e.stopPropagation();
-    setIsAccessKeyHidden(!isAccessKeyHidden);
-  };
-  const [isSecretHidden, setIsSecretHidden] = useState(true);
-  const handleSecretHiddenToggle = e => {
-    e.preventDefault();
-    e.stopPropagation();
-    setIsSecretHidden(!isSecretHidden);
-  };
 
-  const formikHandleChange = (_val, e) => handleChange(e);
-  const formikSetFieldTouched = key => () => setFieldTouched(key, true, true);
-
-  const handleSharedCredsChange = (checked, event) => {
-    setFieldValue('isSharedCred', checked);
-    setIsSharedCred(checked);
-  };
+  // const handleSharedCredsChange = (checked, event) => {
+  //   setFieldValue('isSharedCred', checked);
+  //   setIsSharedCred(checked);
+  // };
 
   const handleProviderChange = option => {
     setFieldValue('volumeSnapshotProvider', option.value);
@@ -123,148 +97,49 @@ const InnerAddEditStorageForm = (props: IOtherProps & FormikProps<IFormValues>) 
     setFieldTouched('volumeSnapshotProvider');
   };
 
+  const DynamicForm = ({ provider }) => {
+    console.log('provider', provider)
+    if (provider === null) {
+      return null;
+    } else {
+      return (
+        <div>
+          {(() => {
+            switch (provider.value) {
+              case 'aws':
+                return <AWSForm {...props} />;
+              // case 'warning':
+              //   return <Warning text={text} />;
+              // case 'error':
+              //   return <Error text={text} />;
+              default:
+                return null;
+            }
+          })()}
+        </div>
+      );
+
+    }
+  }
 
   return (
     <Form onSubmit={handleSubmit} style={{ marginTop: '24px' }}>
-      <FormGroup label="Repository Name" isRequired fieldId={nameKey}>
-        <TextInput
-          onChange={formikHandleChange}
-          onInput={formikSetFieldTouched(nameKey)}
-          onBlur={handleBlur}
-          value={values.name}
-          name={nameKey}
-          type="text"
-          id="storage-name-input"
-          isDisabled={currentStatus.mode === AddEditMode.Edit}
-        />
-        {errors.name && touched.name && (
-          <FormErrorDiv id="feedback-name">{errors.name}</FormErrorDiv>
-        )}
-      </FormGroup>
-      <FormGroup label="S3 Bucket Name" isRequired fieldId={bucketNameKey}>
-        <TextInput
-          onChange={formikHandleChange}
-          onInput={formikSetFieldTouched(bucketNameKey)}
-          onBlur={handleBlur}
-          value={values.bucketName}
-          name={bucketNameKey}
-          type="text"
-          id="storage-bucket-name-input"
-        />
-        {errors.bucketName && touched.bucketName && (
-          <FormErrorDiv id="feedback-bucket-name">{errors.bucketName}</FormErrorDiv>
-        )}
-      </FormGroup>
-      <FormGroup label="S3 Bucket Region" fieldId={bucketRegionKey}>
-        <TextInput
-          onChange={formikHandleChange}
-          onInput={formikSetFieldTouched(bucketRegionKey)}
-          onBlur={handleBlur}
-          value={values.bucketRegion}
-          name={bucketRegionKey}
-          type="text"
-          id="storage-bucket-region-input"
-        />
-        {errors.bucketRegion && touched.bucketRegion && (
-          <FormErrorDiv id="feedback-bucket-name">{errors.bucketName}</FormErrorDiv>
-        )}
-      </FormGroup>
-      <FormGroup label="S3 Endpoint" fieldId={s3UrlKey}>
-        <TextInput
-          onChange={formikHandleChange}
-          onInput={formikSetFieldTouched(s3UrlKey)}
-          onBlur={handleBlur}
-          value={values.s3Url}
-          name={s3UrlKey}
-          type="text"
-          id="storage-s3-url-input"
-        />
-        {errors.s3Url && touched.s3Url && (
-          <FormErrorDiv id="feedback-s3-url">{errors.s3Url}</FormErrorDiv>
-        )}
-      </FormGroup>
-      <FormGroup label="S3 Provider Access Key" isRequired fieldId={accessKeyKey}>
-        <HideWrapper onClick={handleAccessKeyHiddenToggle}>
-          <KeyDisplayIcon id="accessKeyIcon" isHidden={isAccessKeyHidden} />
-        </HideWrapper>
-        <TextInput
-          onChange={formikHandleChange}
-          onInput={formikSetFieldTouched(accessKeyKey)}
-          onBlur={handleBlur}
-          value={values.accessKey}
-          name={accessKeyKey}
-          type={isAccessKeyHidden ? 'password' : 'text'}
-          id="storage-access-key-input"
-        />
-        {errors.accessKey && touched.accessKey && (
-          <FormErrorDiv id="feedback-access-key">{errors.accessKey}</FormErrorDiv>
-        )}
-      </FormGroup>
-      <FormGroup label="S3 Provider Secret Access Key" isRequired fieldId={secretKey}>
-        <HideWrapper onClick={handleSecretHiddenToggle}>
-          <KeyDisplayIcon id="accessKeyIcon" isHidden={isSecretHidden} />
-        </HideWrapper>
-        <TextInput
-          onChange={formikHandleChange}
-          onInput={formikSetFieldTouched(secretKey)}
-          onBlur={handleBlur}
-          value={values.secret}
-          name={secretKey}
-          type={isSecretHidden ? 'password' : 'text'}
-          id="storage-secret-input"
-        />
-        {errors.secret && touched.secret && (
-          <FormErrorDiv id="feedback-secret-key">{errors.secret}</FormErrorDiv>
-        )}
-      </FormGroup>
-      <FormGroup label="VSL Provider Configuration" isRequired fieldId={vslConfigKey}>
-        <Checkbox
-          label="Use BSL credentials for VSL"
-          isChecked={isSharedCred}
-          // onChange={formikHandleChange}
-          onChange={handleSharedCredsChange}
-          aria-label="shared-creds-checkbox"
-          id="shared-creds-checkbox"
-          name="shared-creds-checkbox"
-        />
-        {!isSharedCred &&
-          <Flex flexDirection="column">
-            <Box
-              m="0 0 1em 0 "
-              flex="auto"
-              width={1 / 2}
-            >
-              <Select
-                name="provider"
-                onChange={handleProviderChange}
-                options={providerOptions}
-                value={selectedProvider}
-              />
-            </Box>
-            <Box
-              m="0 0 1em 0 "
-              flex="auto"
-            >
-              <TextArea
-                value={values.vslBlob}
-                onChange={formikHandleChange}
-                aria-label="vsl-blob-text-area"
-                onInput={formikSetFieldTouched(vslConfigKey)}
-                onBlur={handleBlur}
-                // value={values.secret}
-                name={vslBlobKey}
-                type={isSecretHidden ? 'password' : 'text'}
-                id="storage-vsl-blob-input"
+      <Flex flexDirection="column">
+        <Box
+          m="0 0 1em 0 "
+          flex="auto"
+          width={1 / 2}
+        >
+          <Select
+            name="provider"
+            onChange={handleProviderChange}
+            options={providerOptions}
+            value={selectedProvider}
+          />
+        </Box>
+      </Flex>
 
-              />
-              {errors.secret && touched.secret && (
-                <FormErrorDiv id="feedback-secret-key">{errors.secret}</FormErrorDiv>
-              )}
-            </Box>
-          </Flex>
-        }
-
-      </FormGroup>
+      <DynamicForm provider={selectedProvider} {...props} />
 
       <Flex flexDirection="column">
         <Box m="0 0 1em 0 ">
@@ -325,9 +200,10 @@ interface IFormValues {
   accessKey: string;
   secret: string;
   s3Url: string;
-  isSharedSecret?: boolean;
-  vslBlob?: string;
-  volumeSnapshotProvider: string
+  // isSharedSecret?: boolean;
+  // vslBlob?: string;
+  // volumeSnapshotProvider: string
+  bslProvider: string
 }
 interface IOtherProps {
   onAddEditSubmit: any;
@@ -347,9 +223,10 @@ const AddEditStorageForm: any = withFormik({
       accessKey: '',
       secret: '',
       s3Url: '',
-      vslBlob: '',
-      volumeSnapshotProvider: '',
-      isSharedCred: true
+      bslProvider: '',
+      // vslBlob: '',
+      // volumeSnapshotProvider: '',
+      // isSharedCred: true
     };
 
     if (initialStorageValues) {
@@ -359,9 +236,10 @@ const AddEditStorageForm: any = withFormik({
       values.accessKey = initialStorageValues.accessKey || '';
       values.secret = initialStorageValues.secret || '';
       values.s3Url = initialStorageValues.s3Url || '';
-      values.vslBlob = initialStorageValues.vslBlob || '';
-      values.volumeSnapshotProvider = initialStorageValues.volumeSnapshotProvider || '';
-      values.isSharedCred = initialStorageValues.isSharedCred || true;
+      values.bslProvider = initialStorageValues.bslProvider || '';
+      // values.vslBlob = initialStorageValues.vslBlob || '';
+      // values.volumeSnapshotProvider = initialStorageValues.volumeSnapshotProvider || '';
+      // values.isSharedCred = initialStorageValues.isSharedCred || true;
     }
 
     return values;
